# test(boavizt): add tests for client and api