--- a/src/test/java/com/digitalpebble/spruce/modules/boavizta/AbstractBoaviztaTest.java
+++ b/src/test/java/com/digitalpebble/spruce/modules/boavizta/AbstractBoaviztaTest.java
@@ -2,8 +2,8 @@
 
 package com.digitalpebble.spruce.modules.boavizta;
 
-import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.AfterAll;
+import org.junit.jupiter.api.BeforeAll;
 import org.testcontainers.containers.GenericContainer;
 import org.testcontainers.junit.jupiter.Testcontainers;
 import org.testcontainers.utility.DockerImageName;
@@ -13,20 +13,49 @@ public abstract class AbstractBoaviztaTest {
 
     private static final String BOAVIZTAPI_VERSION = "1.3.11";
 
-    protected GenericContainer boaviztapiContainer =
-            new GenericContainer(
-                            DockerImageName.parse(
-                                    "ghcr.io/boavizta/boaviztapi:"+ BOAVIZTAPI_VERSION))
-                    .withExposedPorts(5000);
+    // Static container instance - shared across all test classes
+    private static GenericContainer boaviztapiContainer;
 
-    @BeforeEach
-    void init() {
-        boaviztapiContainer.start();
+    @BeforeAll
+    static void startContainer() {
+        if (boaviztapiContainer == null) {
+            boaviztapiContainer = new GenericContainer(
+                DockerImageName.parse("ghcr.io/boavizta/boaviztapi:" + BOAVIZTAPI_VERSION))
+                .withExposedPorts(5000);
+            
+            boaviztapiContainer.start();
+            
+            // Wait for container to be ready
+            boaviztapiContainer.waitingFor(
+                org.testcontainers.containers.wait.strategy.Wait.forLogMessage(".*Application startup complete.*", 1)
+            );
+        }
     }
 
-    @AfterEach
-    void close() {
-        boaviztapiContainer.close();
+    @AfterAll
+    static void stopContainer() {
+        if (boaviztapiContainer != null) {
+            boaviztapiContainer.stop();
+            boaviztapiContainer = null;
+        }
     }
 
+    // Getter method for test classes that need the container
+    protected static GenericContainer getContainer() {
+        return boaviztapiContainer;
+    }
+
+    // Get the container's host and port for test configuration
+    protected static String getContainerHost() {
+        return boaviztapiContainer.getHost();
+    }
+
+    protected static int getContainerPort() {
+        return boaviztapiContainer.getMappedPort(5000);
+    }
+
+    // Get the full URL for the container
+    protected static String getContainerUrl() {
+        return "http://" + getContainerHost() + ":" + getContainerPort();
+    }
 }