# In this PR
- Added tests around BoaviztAPI module addressing [issue#16](https://github.com/DigitalPebble/spruce/issues/16)
- BoaviztAPIClientTest covers BoaviztAPIClient
- BoaviztAPITest covers BoaviztAPI

```
[INFO] Tests run: 22, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 110.3 s -- in com.digitalpebble.spruce.modules.boavizta.BoaviztAPIClientTest
[INFO] Tests run: 23, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 65.08 s -- in com.digitalpebble.spruce.modules.boavizta.BoaviztAPITest

```

<img width="757" height="157" alt="Screenshot 2025-08-20 at 1 47 51 PM" src="https://github.com/user-attachments/assets/9b6b5a31-9682-484d-8ce7-e1e87aa26b8f" />