@jnioche 2025-08-31
What would you think about changing AbstractBoaviztaTest so that the container is started once and for all and not for every test? This would reduce the time taken by the tests and reduce the risk of getting errors