@nikhiln64 2025-08-31
@jnioche Addressed your comments and here are the improvements after refactoring:
Specific results:
* Previous approach: ~34 seconds for BoaviztAPI tests
* Current approach: 0.092 seconds for BoaviztAPI tests

Total results:


Metric | Before Refactoring | After Refactoring | Improvement
-- | -- | -- | --
BoaviztAPI Tests | ~34s | 0.092s | 370x faster
Total Test Suite | ~40s+ | 3.225s | 12x faster




<img width="726" height="276" alt="Screenshot 2025-08-31 at 8 19 11 AM" src="https://github.com/user-attachments/assets/951ac4f4-7c7c-4139-a657-f6db967a1bf9" />