@jnioche 2025-08-31
thanks @nikhiln64 
Since you have moved the tests to use the mocks, I don't think we need to keep _AbstractBoavizta*_ classes at all, the dependency on org.testcontainers can also be removed. What do you think?