@jnioche 2025-09-02
> thanks @nikhiln64 Since you have moved the tests to use the mocks, I don't think we need to keep _AbstractBoavizta*_ classes at all, the dependency on org.testcontainers can also be removed. What do you think?

I will do that separately.

Thanks for this PR @nikhiln64 and your patience with my comments and suggestions, much appreciated!

Have you used Spruce at all?